FIXED: HTTP and HTTPs
When installer set access to HTTPs via port 443, HTTP access to web
service is denied.  When installer sets access to HTTP, but user
manually configure IIS to use SSL on CMS, HTTPs access to web service is
not found.

Added code to stupidly check if HTTP or HTTPs access to web service is
valid.